stash predmet kalkulacija prosijeka

+++ GradeCalculator/GradeCalculator/Controllers/PredmetController.cs
@@ -69,7 +69,7 @@ namespace GradeCalculator.Controllers
 
             if (subject != null && grades != null)
             {
-                subject.Prosjek = Math.Round(grades.Average(g => g.Vrijednost), 1);
+                subject.Prosjek = Math.Round((double)grades.Average(g => g.Vrijednost), 1);
                 _subjectRepo.Modify(id, subject);
             }
